move embedding size logic to common file
